Move tests folder to top
